% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixoperations.R
\name{adj2el}
\alias{adj2el}
\title{Maps adjacency matrix to edgelist}
\usage{
adj2el(adj, directed = TRUE)
}
\arguments{
\item{adj}{matrix, the adjacency matrix}

\item{directed}{boolean, is the graph directed?}
}
\value{
a dataframe containing the edgelist
}
\description{
Maps adjacency matrix to edgelist
}
