% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilliary.R
\docType{data}
\name{adj_karate}
\alias{adj_karate}
\title{Zachary's Karate Club graph}
\format{a 34x34 matrix}
\source{
package `igraphdata`
}
\usage{
adj_karate
}
\description{
The weighted adjacency reparting interactions among
34 nodes.
}
\keyword{datasets}
