% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/endogenousstats.R
\name{checkGraphtype}
\alias{checkGraphtype}
\title{Check graph input type (for whether it's a graph or a edgelist).}
\usage{
checkGraphtype(graph)
}
\arguments{
\item{graph}{A graph adjacency matrix or an edgelist.}
}
\value{
TRUE or FALSE. Returns TRUE if the provided object \code{graph} is an adjacency matrix.
}
\description{
Returns TRUE if the supplied object \code{graph} is an adjacency matrix. Returns FALSE if the provided object is an edgelist. The function checks whether the edgelist conforms to our standards (sender, target, edgecount).
}
