% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{coef.nrm}
\alias{coef.nrm}
\title{Print method for coeffiients of models of class \code{'nrm'}.}
\usage{
\method{coef}{nrm}(object, ...)
}
\arguments{
\item{object}{object of class \code{'nrm'}.}

\item{\dots}{optional arguments to print methods.}
}
\value{
coefficients of nrm model.
}
\description{
Print method for coeffiients of models of class \code{'nrm'}.
}
\seealso{
\code{\link{nrm}}
}
\author{
Giona Casiraghi
}
