% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{configurationtest}
\alias{configurationtest}
\title{Title}
\usage{
configurationtest(graph, directed, selfloops, nempirical = NULL,
  parallel = NULL)
}
\arguments{
\item{nempirical}{}
}
\description{
Title
}
