% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{contacts.adj}
\alias{contacts.adj}
\title{Highschool contact network adjacency matrix}
\format{
327x327 adjacency matrix
}
\source{
\href{sociopatterns.org}{sociopatters.org}
}
\usage{
data(highschool.predictors)
}
\description{
**contacts.adj**: contains the adjacency matrix of 327 x 327 highschool
students.
}
\references{
Casiraghi, G. Multiplex Network Regression: How do relations
drive interactions? 15 (2017).

Mastrandrea, R., Fournet, J. & Barrat, A. Contact patterns in a high school:
A comparison between data collected using wearable sensors, contact diaries
and friendship surveys. PLoS One 10, 1–26 (2015).
}
\keyword{datasets}
