% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilliary.R
\docType{data}
\name{cospons_mat}
\alias{cospons_mat}
\title{Swiss MPs network adjacency matrix}
\format{163x163 adjacency matrix}
\usage{
data(cospons_mat)
}
\description{
**cospons_mat**: contains the adjacency matrix of 163 x 163 MPs.
}
\keyword{datasets}
