% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{coxsnellR2}
\alias{coxsnellR2}
\title{Computes Cox and Snell pseudo R-squared for nrm models.}
\usage{
coxsnellR2(mod0, mod1, m)
}
\arguments{
\item{mod0}{nrm null model}

\item{mod1}{nrm alternative model}

\item{m}{number of edges}
}
\value{
Cox and Snell pseudo R-squared
}
\description{
Computes Cox and Snell pseudo R-squared for nrm models.
}
\author{
GC
}
