% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPredictors.R
\name{createPredictors.list}
\alias{createPredictors.list}
\title{Create a nrmpredictor object from list}
\usage{
\method{createPredictors}{list}(predictors, ...)
}
\arguments{
\item{predictors}{the dataframe or list of predictors for to apply nrm model selection}
}
\value{
nested list of nrmpredictor class
}
\description{
Create a nrmpredictor object from list
}
\examples{
data('highschool.predictors')
predictors <- createPredictors(highschool.predictors)

}
