% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPredictors.R
\name{createPredictors}
\alias{createPredictors}
\title{Create a nrmpredictor object from passed argument}
\usage{
createPredictors(predictors, ...)
}
\arguments{
\item{predictors}{the dataframe or list of predictors for to apply nrm model selection}

\item{...}{additional parameters passed to the different methods (currently disabled)}
}
\value{
nested list of nrmpredictor class
}
\description{
Create a nrmpredictor object from passed argument
}
\examples{
data('highschool.predictors')
predictors <- createPredictors(highschool.predictors)

}
