% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvhyper.R
\name{dmvhyper_base}
\alias{dmvhyper_base}
\title{Log PMF of multivariate hypergeometric}
\usage{
dmvhyper_base(x, n, k = sum(x), log = FALSE)
}
\arguments{
\item{x}{integer vector of counts drawn in each category}

\item{n}{integer vector of population sizes in each category}

\item{k}{integer total draws (should equal sum(x))}

\item{log}{logical; return log-probability?}
}
\description{
Log PMF of multivariate hypergeometric
}
\keyword{internal}
