% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxilliary.R
\docType{data}
\name{dt}
\alias{dt}
\title{Swiss MPs attribute data frame.}
\format{163x8 data.frame}
\usage{
data(dt)
}
\description{
**dt**: contains different attributes of the 163 MPs, such as their names, 
their party affiliation (variable: *party*), their parliamentary group
affiliation (variable: *parlGroup*), the Canton (or state) they represent
(variable: *canton*), their gender  (variable: *gender*)
and date of birth  (variable: *birthdate*).
}
\keyword{datasets}
