% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dtcommittee}
\alias{dtcommittee}
\title{Swiss MPs committee affiliation data frame.}
\format{163x2 data.frame}
\usage{
data(dtcommittee)
}
\description{
**dtcommittee**: a list of committees each MP was part of during their stay in 
parliament
}
\keyword{datasets}
