% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixoperations.R
\name{el2adj}
\alias{el2adj}
\title{Maps edgelist to adjacency matrix}
\usage{
el2adj(el, nodes = NULL)
}
\arguments{
\item{el}{dataframe containing a (weighted) edgelist. Column 1 is the sender, column 2 is the receiver, column 3 the number of edges.}

\item{nodes}{optional vector containing all node names in case disconnected nodes should be included.}
}
\value{
the (weighted) adjacency matrix corresponding the edgelist passed
}
\description{
Maps edgelist to adjacency matrix
}
