% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{extract.nrm.cluster}
\alias{extract.nrm.cluster}
\title{Extract details from statistical models for table construction. The function has methods for a range of statistical models.}
\usage{
extract.nrm.cluster(model, ...)
}
\arguments{
\item{model}{A statistical model object.}

\item{...}{Custom parameters, which are handed over to subroutines. The arguments are usually passed to the summary function, but in some cases to other functions.}
}
\value{
The function returns a texreg object.
}
\description{
Extract details from statistical models for table construction. The function has methods for a range of statistical models.
}
\author{
L. Brandenberger, G. Casiraghi
}
