% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockmodel.R
\name{fitBlockModel}
\alias{fitBlockModel}
\title{Title}
\usage{
fitBlockModel(adj, labels, directed, selfloops, directedBlocks = FALSE,
  homophily = FALSE, inBlockOnly = FALSE, xi = NULL)
}
\arguments{
\item{selfloops}{}
}
\description{
Title
}
