% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omega.R
\name{FitOmega}
\alias{FitOmega}
\title{Fit propensity matrix for full model}
\usage{
FitOmega(adj, xi, directed, selfloops)
}
\arguments{
\item{adj}{adjacency matrix}

\item{xi}{combinatorial matrix}

\item{directed}{boolean}

\item{selfloops}{boolean}
}
\value{
propensity matrix
}
\description{
(auxiliary function)
}
