% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igraphintegration.R
\name{ghype.igraph}
\alias{ghype.igraph}
\title{Title}
\usage{
\method{ghype}{igraph}(object, directed, selfloops, xi = NULL, omega = NULL,
  unbiased = FALSE, ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
