% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghype.R
\name{ghype.matrix}
\alias{ghype.matrix}
\title{Title}
\usage{
\method{ghype}{matrix}(object, directed, selfloops, xi = NULL,
  omega = NULL, unbiased = FALSE, ...)
}
\arguments{
\item{...}{}
}
\description{
Title
}
