% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{highschool.predictors}
\alias{highschool.predictors}
\title{Highschool contact network predictors}
\format{
5x327x327 list of adjacency matrices
}
\source{
\href{https://sociopatterns.org/}{https://sociopatterns.org/}
}
\usage{
data(highschool.predictors)
}
\description{
**highschool.predictors**: list containing the adjacency matrices
corresponding to the 5 predictors used in Casiraghi2017.
}
\references{
Casiraghi, G. Multiplex Network Regression: How do relations
  drive interactions? 15 (2017).

  Mastrandrea, R., Fournet, J. & Barrat, A. Contact patterns in a high
  school: A comparison between data collected using wearable sensors, contact
  diaries and friendship surveys. PLoS One 10, 1–26 (2015).
}
\keyword{datasets}
