% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockmodel.R
\name{JnBlock}
\alias{JnBlock}
\title{Fisher Information matrix for estimators in block models.}
\usage{
JnBlock(omegaBlocks, xiBlocks, mBlocks, m)
}
\arguments{
\item{omegaBlocks}{the block parameters (vector)}

\item{xiBlocks}{the xi-block (vector)}

\item{mBlocks}{the adj-block (vector)}

\item{m}{the number of edges (scalar)}
}
\value{
Fisher Information matrix
}
\description{
Fisher Information matrix for estimators in block models.
}
