% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{linkSignificance}
\alias{linkSignificance}
\title{Estimate statistical deviations from ghype model}
\usage{
linkSignificance(graph, model, under = FALSE)
}
\arguments{
\item{graph}{an adjacency matrix or a igraph object.}

\item{model}{a ghype model}

\item{under}{boolean, estimate under-represented deviations? Default FALSE.}
}
\value{
matrix of probabilities with same size as adjacency matrix.
}
\description{
linkSignificance allows to estimate the statistical deviations of an observed
graph from a ghype model.
}
