% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logl.R
\name{logl.ghype}
\alias{logl.ghype}
\title{Computes log-likelihood for ghype models from model object}
\usage{
\method{logl}{ghype}(object, adj = NULL, multinomial = NULL, ...)
}
\arguments{
\item{object}{a ghype model}

\item{adj}{an optional adjacency matrix for which to compute the loglikelihood}

\item{multinomial}{optional boolean. Force multinomial approximation?
If not chosen, multinomial chosen for large graphs.}

\item{...}{extra parametes}
}
\value{
loglikelihood value
}
\description{
Computes log-likelihood for ghype models from model object
}
