% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logl.R
\name{logl.matrix}
\alias{logl.matrix}
\title{Computes log-likelihood for ghype models from adjacency.}
\usage{
\method{logl}{matrix}(object, xi, omega, directed, selfloops,
  multinomial = NULL, ...)
}
\arguments{
\item{object}{an adjacency matrix}

\item{xi}{a combinatorial matrix}

\item{omega}{a propensity matrix}

\item{directed}{a boolean argument specifying whether object is directed or not.}

\item{selfloops}{a boolean argument specifying whether the model should incorporate selfloops.}

\item{multinomial}{optional boolean. Force multinomial approximation?
If not chosen, multinomial chosen for large graphs.}

\item{...}{extra parameters}
}
\value{
loglikelihood value
}
\description{
Computes log-likelihood for ghype models from adjacency.
}
