% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logl.R
\name{logl}
\alias{logl}
\title{General method to compute log-likelihood for ghype models.}
\usage{
logl(object, ...)
}
\arguments{
\item{object}{either an adjacency matrix or ghype model
If not chosen, multinomial chosen for large graphs.}

\item{...}{extra parameters}
}
\value{
loglikelihood value
}
\description{
General method to compute log-likelihood for ghype models.
}
