% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logl.R
\name{loglratio}
\alias{loglratio}
\title{Compute log-likelihood ratio for ghype models.}
\usage{
loglratio(mod0, mod1)
}
\arguments{
\item{mod0}{ghype, null model}

\item{mod1}{ghype, alternative model}
}
\value{
scalar, log-likelihood ratio
}
\description{
Compute log-likelihood ratio for ghype models.
}
