% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixoperations.R
\name{mat2vec.ix}
\alias{mat2vec.ix}
\title{Auxiliary function, gives mask for matrix for directed,
undirected etc.}
\usage{
mat2vec.ix(mat, directed, selfloops)
}
\arguments{
\item{mat}{matrix}

\item{directed}{a boolean argument specifying whether object is directed or not.}

\item{selfloops}{a boolean argument specifying whether the model should incorporate selfloops.}
}
\value{
a boolean matrix that can be used to mask adjacency matrices.
}
\description{
Auxiliary function, gives mask for matrix for directed,
undirected etc.
}
