% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{mcfaddenR2}
\alias{mcfaddenR2}
\title{Computes Mc Fadden pseudo R-squared.}
\usage{
mcfaddenR2(
  adj = NULL,
  xi = NULL,
  omega0 = NULL,
  omega1 = NULL,
  directed,
  selfloops,
  mod0 = NULL,
  mod1 = NULL,
  nparam
)
}
\arguments{
\item{adj}{optinal adjacency matrix}

\item{xi}{optional xi matrix}

\item{omega0}{optional propensity matrix of null model}

\item{omega1}{optional propensity matrix of alternative model}

\item{directed}{boolean, is the model directed?}

\item{selfloops}{boolean, are there selfloops?}

\item{mod0}{nrm null model}

\item{mod1}{nrm alternative model}

\item{nparam}{integer, number of parameters}
}
\value{
Mc Fadden pseudo R-squared.
}
\description{
Pass either the models or the model parameters as arguments
}
