% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nr.ci.R
\name{nr.ci}
\alias{nr.ci}
\title{Confidence intervals for nrm models.}
\usage{
nr.ci(nr.m, w, adj, pval)
}
\arguments{
\item{nr.m}{nrm model from which getting coefficients}

\item{w}{list of predictors}

\item{adj}{adjacency matrix}

\item{pval}{numeric. confidence level}
}
\value{
matrix reporting values of predictors and confidence bounds
}
\description{
Internal function to compute confidence intervals for estimated parameters of nrm model
}
