% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nr.significance.R
\name{nr.significance}
\alias{nr.significance}
\title{Computes the significance of more complex model against a simpler model by
means of a likelihood ratio test.}
\usage{
nr.significance(mod0 = NULL, mod1, adj = NULL)
}
\arguments{
\item{mod0}{null nrm model (optional). defaults to the scm model.}

\item{mod1}{alternative nrm model, the model to test}

\item{adj}{adjacency matrix for which performing the test. (optional) defaults to the matrix used for \code{mod1}.}
}
\value{
p-value of the lr test mod0 vs mod1
}
\description{
Computes the significance of more complex model against a simpler model by
means of a likelihood ratio test.
}
