% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nrmSelection.R
\name{nrmSelection}
\alias{nrmSelection}
\alias{nrmSelection.default}
\alias{nrmSelection.nrmpredictor}
\title{Perform AIC forward selection for nrm.}
\usage{
nrmSelection(
  adj,
  predictors,
  directed,
  selfloops,
  pval = 0.05,
  xi = NULL,
  init = NULL,
  ncores = NULL,
  ...
)

\method{nrmSelection}{default}(
  adj,
  predictors,
  directed,
  selfloops,
  pval = 0.05,
  xi = NULL,
  init = NULL,
  ncores = NULL,
  ...
)

\method{nrmSelection}{nrmpredictor}(
  adj,
  predictors,
  directed,
  selfloops,
  pval = 0.05,
  xi = NULL,
  init = NULL,
  ncores = NULL,
  ...
)
}
\arguments{
\item{adj}{the adjacency matrix of the response network}

\item{predictors}{list containing the set of predictors as sublists.}

\item{directed}{logical, is the response network directed?}

\item{selfloops}{logical, do the response network allows selfloops?}

\item{pval}{the significance at which computing confidence intervals.}

\item{xi}{optional, the possibility matrix \eqn{\Xi}.}

\item{init}{optional, initial values passed to the solver to estimate the MLE.}

\item{ncores}{optional, number of cores over which parallelise the task.}

\item{\dots}{optional arguments to method}
}
\value{
A nrm object
}
\description{
Perform AIC forward selection for nrm.
}
\section{Methods (by class)}{
\itemize{
\item \code{default}: Default method for the nrm stepwise selection.

\item \code{nrmpredictor}: Method for the nrm stepwise selection when nrmpredictors are passed.
}}

\examples{
data('highschool.predictors')
nrmSelection(adj=contacts.adj,predictors=highschool.predictors,
  ncores=1,directed=FALSE,selfloops=FALSE)
}
\seealso{
\code{\link{nrm}}
}
\author{
Giona Casiraghi
}
