% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{onlinesim_mat}
\alias{onlinesim_mat}
\title{Swiss MPs committee similarity matrix.}
\format{
163x163 similarity matrix
}
\usage{
data(onlinesim_mat)
}
\description{
**onlinesim_mat**: a similarity matrix of how similar two MPs are in their online
social media presence (shared supportees).
}
\keyword{datasets}
