% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi_matrix.R
\name{phi_element}
\alias{phi_element}
\title{Compute Phi Elementwise}
\usage{
phi_element(edgecount, xi, sum_xi, m)
}
\arguments{
\item{edgecount}{}

\item{xi}{}

\item{sum_xi}{}

\item{m}{}
}
\description{
Compute Phi Elementwise
}
