% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi_matrix.R
\name{phi_element_binomial}
\alias{phi_element_binomial}
\title{Compute Phi Elementwise with binomial distr}
\usage{
phi_element_binomial(edgecount, p, m)
}
\arguments{
\item{edgecount}{}

\item{p}{}

\item{m}{}
}
\value{

}
\description{
Compute Phi Elementwise with binomial distr
}
