% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi_matrix.R
\name{phi_element_hyperg}
\alias{phi_element_hyperg}
\title{Compute Phi Elementwise with hyperg distr}
\usage{
phi_element_hyperg(edgecount, xi, sum_xi, m)
}
\arguments{
\item{edgecount}{}

\item{xi}{}

\item{sum_xi}{}

\item{m}{}
}
\value{

}
\description{
Compute Phi Elementwise with hyperg distr
}
