% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi_matrix.R
\name{phi_element_wallenius}
\alias{phi_element_wallenius}
\title{Compute Phi Elementwise with wallenius distr}
\usage{
phi_element_wallenius(edgecount, xi, sum_xi, oddsratio, m)
}
\arguments{
\item{edgecount}{}

\item{xi}{}

\item{sum_xi}{}

\item{oddsratio}{}

\item{m}{}
}
\value{

}
\description{
Compute Phi Elementwise with wallenius distr
}
