% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi_matrix.R
\name{phi_matrix}
\alias{phi_matrix}
\title{Compute Phi Matrix}
\usage{
phi_matrix(graph, model, lightMemory = FALSE)
}
\arguments{
\item{graph}{n x n adjacency matrix of the graph of which computing the phi matrix}

\item{model}{ghype model against which computing the phi matrix}

\item{lightMemory}{boolean, store model and graph? default FALSE}
}
\value{
n x n phi matrix
}
\description{
Compute Phi Matrix
}
\examples{

data("adj_karate")
model <- scm(graph = adj_karate, directed = FALSE, selfloops = FALSE)
(phi <- phi_matrix(graph = adj_karate, model = model))

}
