% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi_matrix.R
\name{plot.phi_matrix}
\alias{plot.phi_matrix}
\title{Plot method for objects of class phi_matrix}
\usage{
\method{plot}{phi_matrix}(x, ...)
}
\arguments{
\item{x}{an object of class phi_matrix}

\item{...}{arguments to be passed to methods, such as graphical parameters}
}
\description{
Plot method for objects of class phi_matrix
}
\examples{
data("adj_karate")
model <- scm(graph = adj_karate, directed = FALSE, selfloops = FALSE)
phi <- phi_matrix(graph = adj_karate, model = model)
hmcol<-RColorBrewer::brewer.pal(11,"RdBu")
plot(phi, col=hmcol)

}
