% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.ghype}
\alias{print.ghype}
\title{Print method for ghype object.}
\usage{
\method{print}{ghype}(x, suppressCall = FALSE, ...)
}
\arguments{
\item{x}{ghype model}

\item{suppressCall}{boolean, suppress print of the call}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print method for ghype object.
}
