% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.nrm}
\alias{print.nrm}
\title{Print method for elements of class \code{'nrm'}.}
\usage{
\method{print}{nrm}(x, suppressCall = FALSE, ...)
}
\arguments{
\item{x}{object of class \code{'nrm'}}

\item{suppressCall}{logical, indicating whether to print the call that generated x}

\item{\dots}{optional arguments to print or plot methods.}
}
\description{
Print method for elements of class \code{'nrm'}.
}
\seealso{
\code{\link{nrm}}
}
\author{
Giona Casiraghi
}
