% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.nrm.selection}
\alias{print.nrm.selection}
\title{Print method for elements of class \code{'nrm.selection'}.}
\usage{
\method{print}{nrm.selection}(x, ...)
}
\arguments{
\item{x}{object of class \code{'nrm.selection'}.}

\item{\dots}{optional arguments to print or plot methods.}
}
\description{
Print method for elements of class \code{'nrm.selection'}.
}
\seealso{
\code{nrm.selection}
}
\author{
Giona Casiraghi
}
