% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phi_matrix.R
\name{print.phi_matrix}
\alias{print.phi_matrix}
\title{Print method for objects of class phi_matrix}
\usage{
\method{print}{phi_matrix}(x, ...)
}
\arguments{
\item{x}{an object of class phi_matrix}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Print method for objects of class phi_matrix
}
\examples{
data("adj_karate")
model <- scm(graph = adj_karate, directed = FALSE, selfloops = FALSE)
phi <- phi_matrix(graph = adj_karate, model = model)
print(phi)

}
