% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randomsamples.R
\name{rghype}
\alias{rghype}
\title{Generate random realisations from ghype model.}
\usage{
rghype(nsamples, model, m = NULL, multinomial = NULL, seed = NULL)
}
\arguments{
\item{nsamples}{scalar number of realisations}

\item{model}{ghype model}

\item{m}{optional scalar, number of edges to draw}

\item{multinomial}{optional boolean, draw from multinomial?}

\item{seed}{optional scalar, seed for random sampling.}
}
\value{
list of adjacency matrices.
}
\description{
Generate random realisations from ghype model.
}
