% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mvhyper.R
\name{rmvhyper_base}
\alias{rmvhyper_base}
\title{Random samples from multivariate hypergeometric}
\usage{
rmvhyper_base(nn, n, k)
}
\arguments{
\item{nn}{number of samples}

\item{n}{integer vector of population sizes}

\item{k}{integer total draws}
}
\value{
integer matrix with nn rows and length(n) columns
}
\description{
Random samples from multivariate hypergeometric
}
\keyword{internal}
