% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghype.R
\name{scm}
\alias{scm}
\title{Fit the Soft-Configuration Model}
\usage{
scm(object, directed = NULL, selfloops = NULL, ...)
}
\arguments{
\item{object}{either an adjacency matrix or an igraph graph}

\item{directed}{optional boolean, if not specified detected from object}

\item{selfloops}{optional boolean, if not specified detected from object}

\item{...}{additional parameters}
}
\value{
ghype object
}
\description{
scm is wrapper for \link{ghype} that allows to specify a soft-configuration model.
}
