% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{summary.nrm.selection}
\alias{summary.nrm.selection}
\title{Summary method for elements of class \code{'nrm.selection'}.}
\usage{
\method{summary}{nrm.selection}(object, ...)
}
\arguments{
\item{object}{an object of class 'nrm.selection', usually, a result of a call to \code{nrmSelection}.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Summary method for elements of class \code{'nrm.selection'}.
}
