% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{summary.nrm_selection}
\alias{summary.nrm_selection}
\alias{print.summary.nrm_selection}
\title{Summary method for elements of class \code{'nrm_selection'}.}
\usage{
\method{summary}{nrm_selection}(object, ...)

\method{print}{summary.nrm_selection}(x, ...)
}
\arguments{
\item{object}{an object of class 'nrm_selection', usually, a result of a call to \code{nrm_selection}.}

\item{...}{further arguments passed to or from other methods.}

\item{x}{object of class `summary.nrm_selection` returned by [summary.nrm._selection()].}
}
\value{
The function \code{\link{summary.nrm_selection}} computes and
  returns a list of summary statistics of the fitted
  \code{\link{nrm_selection}} model given in \code{object}.
}
\description{
Summary method for elements of class \code{'nrm_selection'}.
}
