% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{vertexlabels}
\alias{vertexlabels}
\title{Zachary's Karate Club vertex faction assignment}
\format{
a 34-vector with the assignment of nodes to faction 1 or 2
}
\source{
package `igraphdata`
}
\usage{
vertexlabels
}
\description{
Vector reporting the assignment of nodes to communities.
}
\keyword{datasets}
