% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixoperations.R
\name{adj2el}
\alias{adj2el}
\title{Maps adjacency matrix to edgelist}
\usage{
adj2el(adj)
}
\arguments{
\item{adj}{the adjacency matrix}
}
\value{
a dataframe containing the edgelist
}
\description{
Maps adjacency matrix to edgelist
}
