% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockmodel.R
\name{bccm}
\alias{bccm}
\title{Title}
\usage{
bccm(adj, labels, directed, selfloops, directedBlocks = FALSE,
  homophily = FALSE, inBlockOnly = FALSE, xi = NULL)
}
\arguments{
\item{xi}{}
}
\description{
Title
}
