% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/computexi.R
\name{ComputeXi}
\alias{ComputeXi}
\title{Auxilliary function. Computes combinatorial matrix.}
\usage{
ComputeXi(adj, directed, selfloops, regular = FALSE)
}
\arguments{
\item{adj}{adjacency matrix}

\item{directed}{boolean}

\item{selfloops}{boolean}

\item{regular}{boolean. Is the combinatorial matrix computed for configuration model or for regular gnp model? default FALSE.}
}
\value{
combinatorial matrix
}
\description{
Combinatorial matrix computed according to soft
configuration model or 'regular' gnp model.
}
