% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blockmodel.R
\name{fitBlockModel}
\alias{fitBlockModel}
\title{Title}
\usage{
fitBlockModel(adj, labels, directed, selfloops)
}
\arguments{
\item{adj}{}

\item{labels}{}

\item{directed}{}

\item{selfloops}{}
}
\description{
Title
}
