% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tests.R
\name{isNetwork}
\alias{isNetwork}
\title{Title}
\usage{
isNetwork(graph, directed, selfloops, Beta = NULL, nempirical = NULL,
  parallel = FALSE, returnBeta = FALSE)
}
\arguments{
\item{selfloops}{}
}
\description{
Title
}
