% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logl.R
\name{logLik.ghype}
\alias{logLik.ghype}
\title{Extract Log-Likelihood}
\usage{
\method{logLik}{ghype}(object, ...)
}
\arguments{
\item{object}{ghype model.}

\item{...}{(ignored) some methods for this generic function require additional arguments.}
}
\value{
Returns an object of class logLik. This is a number with at least one
attribute, "df" (degrees of freedom), giving the number of (estimated) parameters
in the model.
}
\description{
Extract Log-Likelihood
}
