% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.bccm}
\alias{print.bccm}
\title{Print method for elements of class \code{'bccm'}.}
\usage{
\method{print}{bccm}(x, suppressCall = FALSE, ...)
}
\arguments{
\item{x}{object of class \code{'bccm'}}

\item{suppressCall}{logical, indicating whether to print the call that generated x}

\item{\dots}{optional arguments to print or plot methods.}
}
\description{
Print method for elements of class \code{'bccm'}.
}
\examples{
data('adj_karate')
data('vertexlabels')
bcc.model <- bccm(adj_karate, labels=vertexlabels, directed=FALSE, selfloops=FALSE)
print(bcc.model)

}
\seealso{
\code{\link{bccm}}
}
