% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ghype.R
\name{regularm}
\alias{regularm}
\title{Fit the gnm model}
\usage{
regularm(graph, directed = NULL, selfloops = NULL, ...)
}
\arguments{
\item{graph}{either an adjacency matrix or an igraph graph}

\item{directed}{optional boolean, if not specified detected from graph}

\item{selfloops}{optional boolean, if not specified detected from graph}

\item{...}{additional parameters}
}
\value{
ghype object
}
\description{
regularm is wrapper for \link{ghype} that allows to specify a gnm regular model.
i.e. where all entries of the combinatorial matrix Xi are the same.
}
