% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{summary.nrm}
\alias{summary.nrm}
\title{Summary method for elements of class \code{'nrm'}.}
\usage{
\method{summary}{nrm}(object, ...)
}
\arguments{
\item{object}{an object of class 'nrm', usually, a result of a call to \code{nrm}.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Summary method for elements of class \code{'nrm'}.
}
\examples{

}
