% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrixoperations.R
\name{vec2mat}
\alias{vec2mat}
\title{Auxilliary function, produces matrix from vector}
\usage{
vec2mat(vec, directed, selfloops, n)
}
\arguments{
\item{vec}{vector to be put in matrix form}

\item{directed}{a boolean argument specifying whether object is directed or not.}

\item{selfloops}{a boolean argument specifying whether the model should
incorporate selfloops.}

\item{n}{vector. if length(n)==1, n is the number of vertices. If length(n)==3
first element is number of vertices, second and third elements are number of
vertices for row and column of bipartite matrix.}
}
\value{
matrix nxn generated from vector.
}
\description{
Auxilliary function, produces matrix from vector
}
